pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.4",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "boto3>=1.40.36",
    "crypto>=1.4.1",
    "faiss-cpu>=1.12.0",
    "fastapi[standard,standart]>=0.117.1",
    "hf-xet>=1.1.10",
    "isort>=6.0.1",
    "pre-commit>=4.3.0",
    "pyjwt[crypto]>=2.10.1",
    "pylint>=3.3.8",
    "pymupdf>=1.26.4",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
]
